requirements.txt
openai==0.27.2
python-dotenv==1.0.0
sv_ttk==2.4.5
ttkthemes==3.2.2
